pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-ev-charging"
version = "0.1.0"
description = "Home Assistant custom integration: plans EV charging during the cheapest hours of dynamic pricing."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "twarberg", email = "tim@tlw.dk" }]
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-homeassistant-custom-component>=0.13",
    "hypothesis>=6.100",
    "freezegun>=1.5",
    "ruff>=0.5",
    "mypy>=1.10",
]

[tool.setuptools]
packages = []

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "ANN", "PT", "SIM", "RUF"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "B011"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_unreachable = true
disallow_untyped_decorators = false
explicit_package_bases = true
plugins = []

[[tool.mypy.overrides]]
module = "homeassistant.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytest_homeassistant_custom_component.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
testpaths = ["tests"]

[tool.coverage.run]
source = ["custom_components/smart_ev_charging"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 0
